fix: database update when clone operation failed
The database would be updated, next_vm_id would be incremented and user would be added to machine
table, even if the clone operation failed. This commit checks if the clone operation was successful
before trying to updated the database. In case of an internal server error it will notify user of
failure to provision.